Issue 62 Создать реализацию ручки PUT /admin/expense

+++ migrations/0006_add_auth_tokens.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE auth_tokens
 (
     "id"         BIGSERIAL,
-    "user_id"    INTEGER     NOT NULL,
+    "user_id"    BIGINT     NOT NULL,
     "token"      VARCHAR(32) NOT NULL UNIQUE,
     "created_at" TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "end_date"   TIMESTAMP   NOT NULL,
